% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis-glm-function.R
\name{diagglm}
\alias{diagglm}
\title{test whether the generalized linear model assumptions have been violated}
\usage{
diagglm(x, file = "result.txt")
}
\arguments{
\item{x}{an object, which is a result from a linear model}

\item{y}{is the name of text file you want to create (e.g.,'mydiagnosis.txt'). The default is 'result.txt'
Note that the file will be saved at the working directory}
}
\value{
The statistical tests results, outlierTest, durbinWatsonTest, VIF, and gvlma
}
\section{Functions}{
\itemize{
\item \code{diagglm}: for unusual data, nonlinearity, non-normality, non-constant error variance, auto-correlation
}}

